pyproject.toml
[project]
name = "autobloggy"
version = "0.1.0"
description = "A Codex-first, file-based blog drafting harness."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.11.0",
  "PyYAML>=6.0.2",
  "python-pptx>=1.0.2",
]

[project.scripts]
autobloggy = "autobloggy.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.3.5",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autobloggy"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

